Moved parsers to own module
